requirements.txt
opencv-python-headless
numpy
python-telegram-bot==13.7
tflite-runtime
picamera2
